Add +norecurse to dig(1) options - Tests were falsely passing
If I digged @localhost for a domain I had removed from the
configuration it would still exists @localhost and the public
nameservers.

This was because localhost would recurse to the public servers to find
things it couldn't serve itself.

It's also a good idea to instruct the public servers not to recurse so
leave this on for everyone.