Need to set a apt-listchanges@v.nix.is, not just apt-listchanges@localhost
This is because `append_at_myorigin = yes' in `postconf -d' which
means `apt-listchanges' which is what reportbug(1) sends to is
rewritten as apt-listchanges@v.nix.is and not
apt-listchanges@localhost.

Actually probably nothing sends to apt-listchanges@localhost.

We could probably solve a lot of problems and redundancies in postconf
if we rewrote mail to ^$user$ to $user@localhost if the origin of the
mail = localhost.