v-test-etc: run as nobody & sleep before testing on @reboot
We always fail the tests on reboot because they're started before
e.g. apache has a chance to be brought up. Sleep for 5 minutes before
starting them.

The tests were previously run as root/pdns/www-data, but we can run
them all as nobody. This is more secure and self-documenting.