Revert "Don't depend on a webserver to read quotes.yml"
This reverts commit c67b5abe853b45c7b57605601151ffb0e0b12012.

We need to use curl because this script isn't just used on v. This
would have been evident to anyone reading the git log, ARE YOU
PROFESSIONAL OR NOT?

IT'S FUCKING DISTRACTING when I'm trying to cook dinner and I get
V-CRON fail E-Mails because something broke (in this case cat(1) not
being in the PATH). You weren't getting those E-Mails because YOU'RE
NOT PROFESSIONAL.

Do you want be to trash your lights, DO YOU?

Inspired-by: Christian Bale