proftp.conf: revert Tumi's change to add incoming/
Revert the portion of commit f93510110e that changed the permissions
of the FTP server so that you could no longer write in the root.

This broke the uploading of folders, e.g. in GNOME's FTP client
uploading a folder into incoming/ doesn't work, but uploading single
images works.

In Windows's FTP client I couldn't get any sort of uploading to work
at all, but with this reverted this all works again.

We only use the FTP server as a dropbox anyway, and this change broke
that.