bind: remove ARPA mappings & localhost mapping
For the ARPA this did nothing. The actual reverse mapping is
authorative on hetzner and configured via their UI. I think all this
did is when I was running a recurser that this made me feel better.

Ditto the "localhost" mapping. We're not getting that from DNS now,
certainly not our own.